move properties related methods into own context
