+++ tests/acceptance/config/behat.yml
@@ -18,6 +18,7 @@ default:
         - ChecksumContext:
         - FilesVersionsContext:
         - TransferOwnershipContext:
+        - WebDavPropertiesContext:
 
     apiCapabilities:
       paths:
@@ -25,6 +26,7 @@ default:
       contexts:
         - FeatureContext: *common_feature_context_params
         - CapabilitiesContext:
+        - WebDavPropertiesContext:
 
     apiComments:
       paths:
@@ -32,6 +34,7 @@ default:
       contexts:
         - FeatureContext: *common_feature_context_params
         - CommentsContext:
+        - WebDavPropertiesContext:
 
     apiFederation:
       paths:
@@ -39,6 +42,7 @@ default:
       contexts:
         - FeatureContext: *common_feature_context_params
         - FederationContext:
+        - WebDavPropertiesContext:
 
     apiFavorites:
       paths:
@@ -46,18 +50,21 @@ default:
       contexts:
         - FeatureContext: *common_feature_context_params
         - FavoritesContext:
+        - WebDavPropertiesContext:
 
     apiProvisioning-v1:
       paths:
         - '%paths.base%/../features/apiProvisioning-v1'
       contexts:
         - FeatureContext: *common_feature_context_params
+        - WebDavPropertiesContext:
 
     apiProvisioning-v2:
       paths:
         - '%paths.base%/../features/apiProvisioning-v2'
       contexts:
         - FeatureContext: *common_feature_context_params
+        - WebDavPropertiesContext:
 
     apiSharees:
       paths:
@@ -65,6 +72,7 @@ default:
       contexts:
         - FeatureContext: *common_feature_context_params
         - ShareesContext:
+        - WebDavPropertiesContext:
 
     apiShareManagement:
       paths:
@@ -73,6 +81,7 @@ default:
         - FeatureContext: *common_feature_context_params
         - PublicWebDavContext:
         - TrashbinContext:
+        - WebDavPropertiesContext:
 
     apiShareOperations:
       paths:
@@ -81,6 +90,7 @@ default:
         - FeatureContext: *common_feature_context_params
         - PublicWebDavContext:
         - TrashbinContext:
+        - WebDavPropertiesContext:
 
     apiSharingNotifications:
       paths:
@@ -88,6 +98,7 @@ default:
       contexts:
         - FeatureContext: *common_feature_context_params
         - NotificationsCoreContext:
+        - WebDavPropertiesContext:
 
     apiTags:
       paths:
@@ -95,6 +106,7 @@ default:
       contexts:
         - FeatureContext: *common_feature_context_params
         - TagsContext:
+        - WebDavPropertiesContext:
 
     apiTrashbin:
       paths:
@@ -102,6 +114,7 @@ default:
       contexts:
         - FeatureContext: *common_feature_context_params
         - TrashbinContext:
+        - WebDavPropertiesContext:
 
     apiWebdavOperations:
       paths:
@@ -111,6 +124,7 @@ default:
         - LoggingContext:
         - SearchContext:
         - PublicWebDavContext:
+        - WebDavPropertiesContext:
 
     apiWebdavProperties:
       paths:
@@ -118,6 +132,7 @@ default:
       contexts:
         - FeatureContext: *common_feature_context_params
         - LoggingContext:
+        - WebDavPropertiesContext:
 
     cliProvisioning:
       paths:
@@ -126,6 +141,7 @@ default:
         - FeatureContext: *common_feature_context_params
         - EmailContext:
         - OccContext:
+        - WebDavPropertiesContext:
 
     cliMain:
       paths:
@@ -133,6 +149,7 @@ default:
       contexts:
         - FeatureContext: *common_feature_context_params
         - OccContext:
+        - WebDavPropertiesContext:
 
     cliBackground:
       paths:
@@ -140,6 +157,7 @@ default:
       contexts:
         - FeatureContext: *common_feature_context_params
         - OccContext:
+        - WebDavPropertiesContext:
 
     cliTrashbin:
       paths:
@@ -148,6 +166,7 @@ default:
         - FeatureContext: *common_feature_context_params
         - OccContext:
         - TrashbinContext:
+        - WebDavPropertiesContext:
 
     webUIAdminSettings:
       paths:
@@ -165,6 +184,7 @@ default:
         - WebUIGeneralContext:
         - WebUILoginContext:
         - OccContext:
+        - WebDavPropertiesContext:
 
     webUIComments:
       paths:
@@ -175,6 +195,7 @@ default:
         - WebUILoginContext:
         - WebUIFilesContext:
         - WebUISharingContext:
+        - WebDavPropertiesContext:
 
     webUIFavorites:
       paths:
@@ -184,6 +205,7 @@ default:
         - WebUIFilesContext:
         - WebUIGeneralContext:
         - WebUILoginContext:
+        - WebDavPropertiesContext:
 
     webUIFiles:
       paths:
@@ -198,6 +220,7 @@ default:
         - WebUISharingContext:
         - OccContext:
         - PublicWebDavContext:
+        - WebDavPropertiesContext:
 
     webUILogin:
       paths:
@@ -209,6 +232,7 @@ default:
         - WebUIGeneralContext:
         - WebUILoginContext:
         - WebUIPersonalGeneralSettingsContext:
+        - WebDavPropertiesContext:
 
     webUIMoveFilesFolders:
       paths:
@@ -220,6 +244,7 @@ default:
         - WebUILoginContext:
         - WebUISharingContext:
         - PublicWebDavContext:
+        - WebDavPropertiesContext:
 
     webUIPersonalSettings:
       paths:
@@ -234,6 +259,7 @@ default:
         - WebUIPersonalSecuritySettingsContext:
         - WebUIUserContext:
         - OccContext:
+        - WebDavPropertiesContext:
 
     webUIRenameFiles:
       paths:
@@ -245,6 +271,7 @@ default:
         - WebUILoginContext:
         - WebUISharingContext:
         - PublicWebDavContext:
+        - WebDavPropertiesContext:
 
     webUIRenameFolders:
       paths:
@@ -254,6 +281,7 @@ default:
         - WebUIFilesContext:
         - WebUIGeneralContext:
         - WebUILoginContext:
+        - WebDavPropertiesContext:
 
     webUIRestrictSharing:
       paths:
@@ -264,6 +292,7 @@ default:
         - WebUIGeneralContext:
         - WebUILoginContext:
         - WebUISharingContext:
+        - WebDavPropertiesContext:
 
     webUISharingExternal:
       paths:
@@ -277,6 +306,7 @@ default:
         - WebUILoginContext:
         - WebUISharingContext:
         - PublicWebDavContext:
+        - WebDavPropertiesContext:
 
     webUISharingInternalGroups:
       paths:
@@ -288,6 +318,7 @@ default:
         - WebUILoginContext:
         - WebUISharingContext:
         - WebUIAdminSharingSettingsContext:
+        - WebDavPropertiesContext:
 
     webUISharingInternalUsers:
       paths:
@@ -299,6 +330,7 @@ default:
         - WebUILoginContext:
         - WebUISharingContext:
         - WebUIAdminSharingSettingsContext:
+        - WebDavPropertiesContext:
 
     webUISharingNotifications:
       paths:
@@ -311,6 +343,7 @@ default:
         - WebUILoginContext:
         - WebUINotificationsContext:
         - WebUISharingContext:
+        - WebDavPropertiesContext:
 
     webUITags:
       paths:
@@ -323,6 +356,7 @@ default:
         - WebUIFilesContext:
         - WebUISharingContext:
         - WebUITagsContext:
+        - WebDavPropertiesContext:
 
     webUITrashbin:
       paths:
@@ -333,6 +367,7 @@ default:
         - WebUIFilesContext:
         - WebUIGeneralContext:
         - WebUILoginContext:
+        - WebDavPropertiesContext:
 
     webUIUpload:
       paths:
@@ -344,6 +379,7 @@ default:
         - WebUILoginContext:
         - WebUISharingContext:
         - PublicWebDavContext:
+        - WebDavPropertiesContext:
 
     webUIWebdavLocks:
       paths:
@@ -357,6 +393,7 @@ default:
         - WebUIWebDavLockingContext:
         - WebUISharingContext:
         - PublicWebDavContext:
+        - WebDavPropertiesContext:
 
   extensions:
     jarnaiz\JUnitFormatter\JUnitFormatterExtension:
